Description of my process
To start off, I needed to fix my syntax errors from the quiz, which I did and consisted mainly of semi-colon errors. 
I started to continue the logic for the printOrder() function that I had left on. My thought process in making this function
was that I need to first use math to calculate how much space is left, and then use a for-loop to create that many space. I realized
that I needed to use a for-loop because the length of the food in the inventory changed. After making the changes to the previous function, I 
made the printTotal() function that used very simillar logic to the printOrder() logic. However, I realized that, in this case, we knew the maximum length
of the descriptor words, meaning 'total', 'subtotal', and 'tax', which was 8 due to the 'subtotal' being the longest. What I saw that I could do was make the
smaller sized words equal in string length to subtotal by adding spaces in the string to get to 8 total characters. Now since all the descriptors have the same 
length, I just needed to program a constant number of spaces. Lastly, since the prices could be 4 or 5 charactes long (including decimal point), I used condtionals
to format the numbers so that an extra space was made in front of the prices with 4 characters, to make them 5. In general, my process invovled a lot of checking
the alignment of the printer and finding ways to break down the complexity into smaller parts that allowed me to slowly tackle the problem. I struggled with getting
all of the items to align within 40 characters, but the way I figured this out was by calculating the out the spacesby hand and double checking in code.