Description of my process to make this program
The process I used to make the changes from the quiz version to the current version is as follows. 
First, I worked through all the syntax errors that came when I had just type the code directly from the quiz.
They were mainly missing semi-colons and a few times where I needed to declare and initialize a variable, not just declare. 
After those changes were made, I realized that I needed to actually call the printer class within the file, so then I decided to create an instance of the printer class. 
From the quiz the printer class already had a constructor that required arguments, so I passed in the necessary parameters to this printer. 
Before I could call the functions needed from the printer, I needed to actually define them in Printer.java, so I went over on the other file and completed the methods (see other comment for other file comments). 
After completing the methods, I called them from the Printer instance. In general, my thought process in making this file involved lots of cross-checking to make sure that all necessary parameters were recorded and passed in the correct order. 
Also, I had to double-check the user interface for make it look clean multiple times, so user interface design was a large concern of mine in this file.